// frontend/src/models/characters/stat/sub/ISkills.ts
export default interface ISkills {
    athletics: number;
    acrobatics: number;
    sleightHand: number;
    stealth: number;
    arcana: number;
    history: number;
    investigation: number;
    nature: number;
    religion: number;
    animalHandling: number;
    insight: number;
    medicine: number;
    perception: number;
    survival: number;
    deception: number;
    intimidation: number;
    performance: number;
    persuasion: number;
}